Support mtrace() via MALLOC_TRACE
If the shell variable MALLOC_TRACE is set when the snippet is sourced
for a debugging configuration, it is exported and the preprocessor macro
MALLOC_TRACE will be defined when compiling the C program.

main() should then check whether this macro has been defined and call
mtrace() if it is.

The shell variable MALLOC_TRACE must be set to the pathname where the
malloc trace shall be written to when the program is run with mtrace()
enabled.